requirements.txt
psycopg2-binary==2.9.1
django==3.2.5
djangorestframework==3.12.4
gunicorn==19.9.0
